pyproject.toml
[project]
name = "mqtt-manager"
version = "0.1.0"
description = "PySide6 desktop application for managing a Mosquitto MQTT broker on Raspberry Pi"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6",
    "paramiko>=3.4",
    "paho-mqtt>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
mqtt-manager = "mqtt_manager.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mqtt_manager"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
